product-store.js aufgeräumt
