--- src/pages/shop/store/products/getters.js
@@ -1,12 +1,0 @@
-const getters = {
-  getAll(state) {
-    return state.all
-  },
-  getCurrentProduct(state) {
-    return state.currentId
-      ? state.products[state.currentId]
-      : {}
-  }
-}
-
-export default getters
