--- src/pages/shop/store/products/mutations.js
@@ -1,10 +1,0 @@
-export default {
-  setProducts(state, products) {
-    state.all = products
-  },
-
-  /* decrementProductInventory(state, {id}) {
-     const product = state.all.find(product => product.id === id)
-     product.inventory--
-   }*/
-}
