+++ src/plugins/index.js
@@ -1,14 +1,14 @@
 /**
- * plugins/index.js
+ * plugins/product-store.js
  *
  * Automatically included in `./src/main.js`
  */
 
 // Plugins
-import { loadFonts } from './webfontloader'
+import {loadFonts} from './webfontloader'
 import vuetify from './vuetify'
 
-export function registerPlugins (app) {
+export function registerPlugins(app) {
   loadFonts()
   app.use(vuetify)
 }
